+++ src/backend/people/settings.py
@@ -547,6 +547,13 @@ class Base(Configuration):
         environ_prefix=None,
     )
 
+    # SUPPORT
+    CRISP_WEBSITE_ID = values.Value(
+        default=None,
+        environ_name="CRISP_WEBSITE_ID",
+        environ_prefix=None,
+    )
+
     # MAILBOX-PROVISIONING API
     WEBMAIL_URL = values.Value(
         default=None,